refactor: moved chart folder and status card folder to components
