+++ src/components/Manage/StatusCard/styles.ts
